Add template expand action
While this particular action is in an especially poor state at this time
I want to go ahead and commit it. It does work though I believe that
behavior should be improved. Moreover all of it should be tested and
thus I will focus the next PR on testing for plugins.