Add template expansion plugin
* Update demo script to use templates

* Add template expand action

While this particular action is in an especially poor state at this time
I want to go ahead and commit it. It does work though I believe that
behavior should be improved. Moreover all of it should be tested and
thus I will focus the next PR on testing for plugins.

* Add integration test

This demo file exercises the template feature along with a sample that's
inside of the plugin directory. Will be expanding on this directory
structure now where we will have a plugin directory for all of the
non-core code.

* Fix golanglint-ci version and cache in github actions 

The older version of golanglint-ci (like used in their example) is only useful when you need to pin older versions. Removed the explicit version selection. Also added the cache keys to deal with potential version issues in deps. Just in case we use a separate key for each build in the matrix, but more importantly we hash the module lock files.