Update targets in Makefile
The essence of change was updating targets, so that they work with the
new structure of the repo moving the main package under cmd. But we also
fine-tuned the targets to make them more useful and flexible - adding
variables for a bin output directory we can use to test the build along
with install targets.