Refactor/move cli (#6)
* Move main to cmd package

* Move main package under cmd/carbon

Long-term this can make more sense as ancillary tools and types of UI
may be added. Short run it requires less fiddling with `go install` to
get the binary named how I want in GOBIN.

* Update targets in Makefile

The essence of change was updating targets, so that they work with the
new structure of the repo moving the main package under cmd. But we also
fine-tuned the targets to make them more useful and flexible - adding
variables for a bin output directory we can use to test the build along
with install targets.

* Update build descriptor to reflect dirty State

This change required that we use a new git command to calculate the
version and we also added a dirty state. To use dirty state effectively
when we're not committing our vendor deps required adding the dir to the gitignore.

* Use Go 1.18